refactor(sglang): move checkpoint adapter into checkpoint_restore.py
Move SGLangCheckpointAdapter, CHECKPOINT_SLEEP_MODE_LEVEL,
CHECKPOINT_MEMORY_TAGS, and setup_engine_for_checkpoint from main.py
into checkpoint_restore.py. main.py now imports from checkpoint_restore
the same way vLLM does, keeping all checkpoint/restore logic in one
place per framework.
